+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed import lock
+- Disabled folder watch
+
 ## [0.0.35] - 2020-12-16
 
 ### Changed